# Move `QuantumCircuit.assign_parameters` to Rust