--- a/crates/circuit/src/imports.rs
+++ b/crates/circuit/src/imports.rs
@@ -45,6 +45,8 @@ impl ImportOnceCell {
                 .unwrap()
                 .getattr(self.object)
                 .unwrap()
+                .downcast_into()
+                .unwrap()
                 .unbind()
         })
     }
@@ -76,6 +78,7 @@ pub static CONTROLLED_GATE: ImportOnceCell =
 pub static DEEPCOPY: ImportOnceCell = ImportOnceCell::new("copy", "deepcopy");
 pub static QI_OPERATOR: ImportOnceCell = ImportOnceCell::new("qiskit.quantum_info", "Operator");
 pub static WARNINGS_WARN: ImportOnceCell = ImportOnceCell::new("warnings", "warn");
+pub static UUID: ImportOnceCell = ImportOnceCell::new("uuid", "UUID");
 
 /// A mapping from the enum variant in crate::operations::StandardGate to the python
 /// module path and class name to import it. This is used to populate the conversion table