--- a/crates/circuit/src/imports.rs
+++ b/crates/circuit/src/imports.rs
@@ -45,8 +45,6 @@ impl ImportOnceCell {
                 .unwrap()
                 .getattr(self.object)
                 .unwrap()
-                .downcast_into()
-                .unwrap()
                 .unbind()
         })
     }