--- a/qiskit/circuit/_utils.py
+++ b/qiskit/circuit/_utils.py
@@ -19,19 +19,6 @@ import numpy
 from qiskit import _numpy_compat
 from qiskit.exceptions import QiskitError
 from qiskit.circuit.exceptions import CircuitError
-from .parametervector import ParameterVectorElement
-
-
-def sort_parameters(parameters):
-    """Sort an iterable of :class:`.Parameter` instances into a canonical order, respecting the
-    ordering relationships between elements of :class:`.ParameterVector`\\ s."""
-
-    def key(parameter):
-        if isinstance(parameter, ParameterVectorElement):
-            return (parameter.vector.name, parameter.index)
-        return (parameter.name,)
-
-    return sorted(parameters, key=key)
 
 
 def _compute_control_matrix(base_mat, num_ctrl_qubits, ctrl_state=None):