--- a/qiskit/circuit/quantumcircuit.py
+++ b/qiskit/circuit/quantumcircuit.py
@@ -15,6 +15,8 @@
 """Quantum circuit object."""
 
 from __future__ import annotations
+
+import collections.abc
 import copy as _copy
 import itertools
 import multiprocessing as mp
@@ -45,7 +47,6 @@ from qiskit.circuit.gate import Gate
 from qiskit.circuit.parameter import Parameter
 from qiskit.circuit.exceptions import CircuitError
 from . import _classical_resource_map
-from ._utils import sort_parameters
 from .controlflow import ControlFlowOp, _builder_utils
 from .controlflow.builder import CircuitScopeInterface, ControlFlowBuilderBlock
 from .controlflow.break_loop import BreakLoopOp, BreakLoopPlaceholder
@@ -1126,9 +1127,6 @@ class QuantumCircuit:
         self._calibrations: DefaultDict[str, dict[tuple, Any]] = defaultdict(dict)
         self.add_register(*regs)
 
-        # Cache to avoid re-sorting parameters
-        self._parameters = None
-
         self._layout = None
         self.global_phase = global_phase
 
@@ -1265,7 +1263,6 @@ class QuantumCircuit:
         else:
             data_input = list(data_input)
         self._data.clear()
-        self._parameters = None
         # Repopulate the parameter table with any global-phase entries.
         self.global_phase = self.global_phase
         if not data_input:
@@ -2491,9 +2488,7 @@ class QuantumCircuit:
         :meta public:
         """
         if _standard_gate:
-            new_param = self._data.append(instruction)
-            if new_param:
-                self._parameters = None
+            self._data.append(instruction)
             self.duration = None
             self.unit = "dt"
             return instruction
@@ -2505,19 +2500,16 @@ class QuantumCircuit:
         # instruction param the inner rust append method will raise a runtime error.
         # When this happens we need to handle the parameters separately.
         # This shouldn't happen in practice but 2 tests were doing this and it's not
-        # explicitly prohibted by the API so this and the `params` optional argument
-        # path guard against it.
+        # explicitly prohibted by the API.
         try:
-            new_param = self._data.append(instruction)
+            self._data.append(instruction)
         except RuntimeError:
-            params = []
-            for idx, param in enumerate(instruction.operation.params):
-                if isinstance(param, (ParameterExpression, QuantumCircuit)):
-                    params.append((idx, list(set(param.parameters))))
-            new_param = self._data.append(instruction, params)
-        if new_param:
-            # clear cache if new parameter is added
-            self._parameters = None
+            params = [
+                (idx, param.parameters)
+                for idx, param in enumerate(instruction.operation.params)
+                if isinstance(param, (ParameterExpression, QuantumCircuit))
+            ]
+            self._data.append_manual_params(instruction, params)
 
         # Invalidate whole circuit duration if an instruction is added
         self.duration = None
@@ -2573,7 +2565,7 @@ class QuantumCircuit:
                 A similar method, but for :class:`.expr.Var` run-time variables instead of
                 :class:`.Parameter` compile-time parameters.
         """
-        if (parameter := self._data.get_param_from_name(name)) is None:
+        if (parameter := self._data.get_parameter_by_name(name)) is None:
             if default is Ellipsis:
                 raise KeyError(f"no parameter named '{name}' is present")
             return default
@@ -3645,8 +3637,6 @@ class QuantumCircuit:
         cpy._data = CircuitData(
             self._data.qubits, self._data.clbits, global_phase=self._data.global_phase
         )
-        # Invalidate parameters caching.
-        cpy._parameters = None
 
         cpy._calibrations = _copy.deepcopy(self._calibrations)
         cpy._metadata = _copy.deepcopy(self._metadata)
@@ -4033,16 +4023,6 @@ class QuantumCircuit:
         Args:
             angle (float, ParameterExpression): radians
         """
-        # If we're currently parametric, we need to throw away the references.  This setter is
-        # called by some subclasses before the inner `_global_phase` is initialized.
-        if isinstance(getattr(self._data, "global_phase", None), ParameterExpression):
-            self._parameters = None
-        if isinstance(angle, ParameterExpression):
-            if angle.parameters:
-                self._parameters = None
-        else:
-            angle = _normalize_global_phase(angle)
-
         if self._control_flow_scopes:
             self._control_flow_scopes[-1].global_phase = angle
         else:
@@ -4107,11 +4087,8 @@ class QuantumCircuit:
         Returns:
             The sorted :class:`.Parameter` objects in the circuit.
         """
-        # parameters from gates
-        if self._parameters is None:
-            self._parameters = sort_parameters(self._unsorted_parameters())
         # return as parameter view, which implements the set and list interface
-        return ParameterView(self._parameters)
+        return ParameterView(self._data.parameters)
 
     @property
     def num_parameters(self) -> int:
@@ -4129,7 +4106,7 @@ class QuantumCircuit:
         """
         # This should be free, by accessing the actual backing data structure of the table, but that
         # means that we need to copy it if adding keys from the global phase.
-        return self._data.get_params_unsorted()
+        return self._data.unsorted_parameters()
 
     @overload
     def assign_parameters(
@@ -4242,108 +4219,28 @@ class QuantumCircuit:
         if inplace:
             target = self
         else:
+            if not isinstance(parameters, dict):
+                # We're going to need to access the sorted order wihin the inner Rust method on
+                # `target`, so warm up our own cache first so that subsequent calls to
+                # `assign_parameters` on `self` benefit as well.
+                _ = self._data.parameters
             target = self.copy()
             target._increment_instances()
             target._name_update()
 
-        # Normalize the inputs into simple abstract interfaces, so we've dispatched the "iteration"
-        # logic in one place at the start of the function.  This lets us do things like calculate
-        # and cache expensive properties for (e.g.) the sequence format only if they're used; for
-        # many large, close-to-hardware circuits, we won't need the extra handling for
-        # `global_phase` or recursive definition binding.
-        #
-        # During normalisation, be sure to reference 'parameters' and related things from 'self' not
-        # 'target' so we can take advantage of any caching we might be doing.
-        if isinstance(parameters, dict):
+        if isinstance(parameters, collections.abc.Mapping):
             raw_mapping = parameters if flat_input else self._unroll_param_dict(parameters)
-            # Remember that we _must not_ mutate the output of `_unsorted_parameters`.
-            our_parameters = self._unsorted_parameters()
+            our_parameters = self._data.unsorted_parameters()
             if strict and (extras := raw_mapping.keys() - our_parameters):
                 raise CircuitError(
                     f"Cannot bind parameters ({', '.join(str(x) for x in extras)}) not present in"
                     " the circuit."
                 )
             parameter_binds = _ParameterBindsDict(raw_mapping, our_parameters)
+            target._data.assign_parameters_mapping(parameter_binds)
         else:
-            our_parameters = self.parameters
-            if len(parameters) != len(our_parameters):
-                raise ValueError(
-                    "Mismatching number of values and parameters. For partial binding "
-                    "please pass a dictionary of {parameter: value} pairs."
-                )
-            parameter_binds = _ParameterBindsSequence(our_parameters, parameters)
-
-        # Clear out the parameter table for the relevant entries, since we'll be binding those.
-        # Any new references to parameters are reinserted as part of the bind.
-        target._parameters = None
-        # This is deliberately eager, because we want the side effect of clearing the table.
-        all_references = [
-            (parameter, value, target._data.pop_param(parameter.uuid.int, parameter.name, ()))
-            for parameter, value in parameter_binds.items()
-        ]
-        seen_operations = {}
-        # The meat of the actual binding for regular operations.
-        for to_bind, bound_value, references in all_references:
-            update_parameters = (
-                tuple(bound_value.parameters)
-                if isinstance(bound_value, ParameterExpression)
-                else ()
-            )
-            for inst_index, index in references:
-                if inst_index == self._data.global_phase_param_index:
-                    operation = None
-                    seen_operations[inst_index] = None
-                    assignee = target.global_phase
-                    validate = _normalize_global_phase
-                else:
-                    operation = target._data[inst_index].operation
-                    seen_operations[inst_index] = operation
-                    assignee = operation.params[index]
-                    validate = operation.validate_parameter
-                if isinstance(assignee, ParameterExpression):
-                    new_parameter = assignee.assign(to_bind, bound_value)
-                    for parameter in update_parameters:
-                        if not target._data.contains_param(parameter.uuid.int):
-                            target._data.add_new_parameter(parameter, inst_index, index)
-                        else:
-                            target._data.update_parameter_entry(
-                                parameter.uuid.int,
-                                inst_index,
-                                index,
-                            )
-                    if not new_parameter.parameters:
-                        new_parameter = validate(new_parameter.numeric())
-                elif isinstance(assignee, QuantumCircuit):
-                    new_parameter = assignee.assign_parameters(
-                        {to_bind: bound_value}, inplace=False, flat_input=True
-                    )
-                else:
-                    raise RuntimeError(  # pragma: no cover
-                        f"Saw an unknown type during symbolic binding: {assignee}."
-                        " This may indicate an internal logic error in symbol tracking."
-                    )
-                if inst_index == self._data.global_phase_param_index:
-                    # We've already handled parameter table updates in bulk, so we need to skip the
-                    # public setter trying to do it again.
-                    target._data.global_phase = new_parameter
-                else:
-                    temp_params = operation.params
-                    temp_params[index] = new_parameter
-                    operation.params = temp_params
-                    target._data.setitem_no_param_table_update(
-                        inst_index,
-                        target._data[inst_index].replace(operation=operation, params=temp_params),
-                    )
-
-        # After we've been through everything at the top level, make a single visit to each
-        # operation we've seen, rebinding its definition if necessary.
-        for operation in seen_operations.values():
-            if (
-                definition := getattr(operation, "_definition", None)
-            ) is not None and definition.num_parameters:
-                definition.assign_parameters(
-                    parameter_binds.mapping, inplace=True, flat_input=True, strict=False
-                )
+            parameter_binds = _ParameterBindsSequence(target._data.parameters, parameters)
+            target._data.assign_parameters_sequence(parameters)
 
         # Finally, assign the parameters inside any of the calibrations.  We don't track these in
         # the `ParameterTable`, so we manually reconstruct things.
@@ -4380,7 +4277,6 @@ class QuantumCircuit:
                 for gate, calibrations in target._calibrations.items()
             ),
         )
-        target._parameters = None
         return None if inplace else target
 
     def _unroll_param_dict(
@@ -5975,7 +5871,6 @@ class QuantumCircuit:
         if not self._data:
             raise CircuitError("This circuit contains no instructions.")
         instruction = self._data.pop()
-        self._parameters = None
         return instruction
 
     @typing.overload
@@ -6629,7 +6524,6 @@ class _OuterCircuitScopeInterface(CircuitScopeInterface):
 
     def extend(self, data: CircuitData):
         self.circuit._data.extend(data)
-        self.circuit._parameters = None
         self.circuit.duration = None
         self.circuit.unit = "dt"
 
@@ -6788,11 +6682,3 @@ def _bit_argument_conversion_scalar(specifier, bit_sequence, bit_set, type_):
         else f"Invalid bit index: '{specifier}' of type '{type(specifier)}'"
     )
     raise CircuitError(message)
-
-
-def _normalize_global_phase(angle):
-    """Return the normalized form of an angle for use in the global phase.  This coerces to float if
-    possible, and fixes to the interval :math:`[0, 2\\pi)`."""
-    if isinstance(angle, ParameterExpression) and angle.parameters:
-        return angle
-    return float(angle) % (2.0 * np.pi)