--- a/test/python/circuit/library/test_blueprintcircuit.py
+++ b/test/python/circuit/library/test_blueprintcircuit.py
@@ -77,17 +77,17 @@ class TestBlueprintCircuit(QiskitTestCase):
 
         with self.subTest(msg="after building"):
             self.assertGreater(len(mock._data), 0)
-            self.assertEqual(mock._data.num_params(), 1)
+            self.assertEqual(mock._data.num_parameters(), 1)
 
         mock._invalidate()
         with self.subTest(msg="after invalidating"):
             self.assertFalse(mock._is_built)
-            self.assertEqual(mock._data.num_params(), 0)
+            self.assertEqual(mock._data.num_parameters(), 0)
 
         mock._build()
         with self.subTest(msg="after re-building"):
             self.assertGreater(len(mock._data), 0)
-            self.assertEqual(mock._data.num_params(), 1)
+            self.assertEqual(mock._data.num_parameters(), 1)
 
     def test_calling_attributes_works(self):
         """Test that the circuit is constructed when attributes are called."""