@mtreinish 2024-07-26
The example Steve shared was:
```python
from qiskit.circuit import QuantumCircuit, Parameter, ClassicalRegister, QuantumRegister

from qiskit.transpiler.passes import TranslateParameterizedGates

SUPPORTED_GATES = [
    "rx",
    "ry",
    "rz",
    "rzx",
    "rzz",
    "ryy",
    "rxx",
    "cx",
    "cy",
    "cz",
    "ccx",
    "swap",
    "iswap",
    "h",
    "t",
    "s",
    "sdg",
    "x",
    "y",
    "z",
]

a = Parameter("a")
qc = QuantumCircuit(1)
qc.h(0)
qc.p(a, 0)
qc.h(0)

print(qc)
print(qc.parameters)

translator = TranslateParameterizedGates(SUPPORTED_GATES)
qc = translator(qc)

print(qc)
print(qc.parameters)


qc1 = qc.copy()
qr_aux = QuantumRegister(1, "qr_aux")
cr_aux = ClassicalRegister(1, "cr_aux")
qc1.add_register(qr_aux)
qc1.add_register(cr_aux)
qc1.h(qr_aux)
qc1.data.insert(0, qc1.data.pop()) # This line seems to mess things up

print(qc1)
print(qc1.parameters)

cct = qc1.assign_parameters({a: 0.44})

print(cct)
print(cct.parameters)
```

on main it looked like the final circuit printed was not binding `a` on the correct index in `CircuitData` in the rust code. So `Parameter('a')`  was still in the `params` field on the first `PackedInstruction`/`CircuitInstruction` but the parameter table didn't have it anymore and said it was bound. I assume it was something around the `pop()` or `insert()` where there was an edge case in the tracking/reindexing of the parameter table on main. But this fixes it as part of your larger refactoring. We probably should add a test that does something like this to make sure we don't regress.