@jakelishman 2024-07-26
Yeah, I'm pretty sure I know exactly where that bug had come in, and I remember moving a line of code that I imagine is what fixed it. In the insert code, I think the new data is written in _before_ we called untrack on it, so it untracked the wrong thing.