@jakelishman 2024-07-31
I added a test that catches Steve's case in 1cd8d08.  It turned out to be a bit more complex than I'd originally thought - it's a particular interaction with the parametric global-phase tracking in addition to the pop/insert retracking required, and somewhere along the line, that was causing the mess-up.