@jakelishman 2024-07-31
Ok, comments should be addressed, and the `AnnotatedOperation` problems should be fixed by 83da2f4662.

Fwiw they've been pre-existing in some form or another ever since `AnnotatedOperation` was first introduced.